auth_system/requirements.txt

#install by running: pip install -r requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
argon2-cffi
pydantic

